New rule 0022 GlobalLanguage() must be implemented through the "Translation Helper" codeunit
