+++ BusinessCentral.LinterCop.csproj
@@ -112,7 +112,7 @@
     <Compile Include="Design\Rule0018NoEventsInInternalCodeunits.cs" />
     <Compile Include="Design\Rule0019DataClassificationFieldEqualsTable.cs" />
     <Compile Include="Design\Rule0020ApplicationAreaEqualsToPage.cs" />
-    <Compile Include="Design\Rule0021ConfirmImplementConfirmManagement.cs" />
+    <Compile Include="Design\Rule0021BuiltInMethodImplementThroughCodeunit.cs" />
     <Compile Include="Helpers\LinterSettings.cs" />
     <Compile Include="LinterCopAnalyzers.Designer.cs" />
     <Compile Include="LinterCopAnalyzers.Generated.cs" />
