+++ LinterCopAnalyzers.resx
@@ -303,4 +303,7 @@
   <data name="Rule0021ConfirmImplementConfirmManagement" xml:space="preserve">
     <value>Confirm() must be implemented through the "Confirm Management" codeunit from the System Application.</value>
   </data>
+  <data name="Rule0022GlobalLanguageImplementTranslationHelper" xml:space="preserve">
+    <value>GlobalLanguage() must be implemented through the "Translation Helper" codeunit from the Base Application.</value>
+  </data>
 </root>
